+++ examples/document.rs
@@ -3,9 +3,9 @@
 
 use anyhow::Error;
 
-use arangors::{document::options::InsertOptions, Collection, Connection};
+use arangors_lite::{document::options::InsertOptions, Collection, Connection};
 
-use arangors::document::{
+use arangors_lite::document::{
     options::{RemoveOptions, ReplaceOptions, UpdateOptions},
     response::DocumentResponse,
 };
@@ -21,9 +21,8 @@ struct User {
     email: String,
 }
 
-#[cfg_attr(feature = "reqwest_async", tokio::main)]
-#[cfg_attr(feature = "surf_async", async_std::main)]
-#[cfg_attr(feature = "reqwest_blocking", maybe_async::must_be_sync)]
+#[cfg_attr(not(feature = "blocking"), tokio::main)]
+#[cfg_attr(feature = "blocking", maybe_async::must_be_sync)]
 async fn main() -> Result<(), Error> {
     let collection_name = "test_collection_document_example";
 
@@ -146,10 +145,3 @@ async fn main() -> Result<(), Error> {
 
     Ok(())
 }
-
-#[cfg(not(any(
-    feature = "reqwest_blocking",
-    feature = "reqwest_async",
-    feature = "surf_async"
-)))]
-fn main() {}