+++ examples/reqwest_rustls/src/main.rs
@@ -3,10 +3,10 @@
 //! 1. use vanilla arangors without any http client implementation by disabling
 //! `reqwest_async`, `reqwest_blocking` and `surf_async` on arangors.
 //! 2. implement custom reqwest client and enable `rustls` feature gate like in `src/client.rs`.
-//! 3. use custom client with `arangors::GenericConnection`.
+//! 3. use custom client with `arangors_lite::GenericConnection`.
 mod client;
 
-use arangors::GenericConnection;
+use arangors_lite::GenericConnection;
 
 use self::client::ReqwestClient;