+++ tests/aql.rs
@@ -4,7 +4,7 @@
 use pretty_assertions::assert_eq;
 use serde::{Deserialize, Serialize};
 
-use arangors::{AqlQuery, Connection, Document};
+use arangors_lite::{AqlQuery, Connection, Document};
 use common::{connection, test_setup};
 
 use crate::common::{get_arangodb_host, get_root_password, get_root_user};
@@ -17,11 +17,7 @@ struct User {
     pub password: String,
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_aql_str() {
     test_setup();
     let conn = connection().await;
@@ -34,11 +30,7 @@ async fn test_aql_str() {
     assert_eq!(result[0].document.password, "test2_pwd");
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_aql() {
     test_setup();
     let conn = connection().await;
@@ -51,11 +43,7 @@ async fn test_aql() {
     assert_eq!(result[0].document.password, "test2_pwd");
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_aql_bind_vars() {
     test_setup();
     let conn = connection().await;
@@ -70,11 +58,7 @@ async fn test_aql_bind_vars() {
     assert_eq!(result[0].document.password, "test2_pwd");
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_aql_try_bind() {
     test_setup();
     let conn = connection().await;