+++ tests/connection.rs
@@ -1,9 +1,8 @@
 #![allow(unused_imports)]
 #![allow(unused_parens)]
 use pretty_assertions::assert_eq;
-use uclient::ClientExt;
 
-use arangors::{connection::Permission, Connection};
+use arangors_lite::{connection::Permission, Connection};
 use common::{
     connection, get_arangodb_host, get_normal_password, get_normal_user, test_root_and_normal,
     test_setup,
@@ -11,11 +10,7 @@ use common::{
 
 pub mod common;
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_list_databases() {
     test_setup();
     let conn = connection().await;
@@ -25,17 +20,13 @@ async fn test_list_databases() {
     let db_permission = dbs.get("test_db").unwrap();
     match db_permission {
         Permission::ReadOnly | Permission::NoAccess => {
-            assert!(false, "Invalid permission {:?}", db_permission)
+            panic!("Invalid permission {:?}", db_permission)
         }
         _ => {}
     };
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_get_url() {
     test_setup();
     let host = get_arangodb_host();
@@ -49,11 +40,7 @@ async fn test_get_url() {
     assert_eq!(url, host)
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_get_database() {
     test_setup();
     let conn = connection().await;
@@ -63,11 +50,7 @@ async fn test_get_database() {
     assert_eq!(database.is_err(), true);
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_basic_auth() {
     test_setup();
     let host = get_arangodb_host();
@@ -83,11 +66,7 @@ async fn test_basic_auth() {
     assert_eq!(headers.get("Server").unwrap(), "ArangoDB");
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_jwt() {
     test_setup();
     #[maybe_async::maybe_async]