+++ tests/database.rs
@@ -3,7 +3,7 @@
 use log::trace;
 use pretty_assertions::assert_eq;
 
-use arangors::Connection;
+use arangors_lite::Connection;
 use common::{
     connection, get_arangodb_host, get_normal_password, get_normal_user, get_root_password,
     get_root_user, test_root_and_normal, test_setup,
@@ -13,11 +13,7 @@ pub mod common;
 
 const NEW_DB_NAME: &str = "example";
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_create_and_drop_database() {
     test_setup();
     let host = get_arangodb_host();
@@ -30,24 +26,20 @@ async fn test_create_and_drop_database() {
 
     let result = conn.create_database(NEW_DB_NAME).await;
     if let Err(e) = result {
-        assert!(false, "Fail to create database: {:?}", e)
+        panic!("Fail to create database: {:?}", e)
     };
     let result = conn.db(NEW_DB_NAME).await;
     assert_eq!(result.is_err(), false);
 
     let result = conn.drop_database(NEW_DB_NAME).await;
     if let Err(e) = result {
-        assert!(false, "Fail to drop database: {:?}", e)
+        panic!("Fail to drop database: {:?}", e)
     };
     let result = conn.db(NEW_DB_NAME).await;
     assert_eq!(result.is_err(), true);
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_fetch_current_database_info() {
     test_setup();
 
@@ -64,17 +56,13 @@ async fn test_fetch_current_database_info() {
                 trace!("{:?}", info);
                 assert_eq!(info.is_system, false)
             }
-            Err(e) => assert!(false, "Fail to fetch database: {:?}", e),
+            Err(e) => panic!("Fail to fetch database: {:?}", e),
         }
     }
     test_root_and_normal(fetch_current_database).await;
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_get_version() {
     test_setup();
     let conn = connection().await;