+++ tests/document.rs
@@ -5,7 +5,7 @@ use log::trace;
 use pretty_assertions::assert_eq;
 use serde_json::{json, Value};
 
-use arangors::{
+use arangors_lite::{
     document::{
         options::{
             InsertOptions, OverwriteMode, ReadOptions, RemoveOptions, ReplaceOptions, UpdateOptions,
@@ -22,11 +22,7 @@ use std::{convert::TryInto, ptr::null};
 pub mod common;
 
 #[cfg(not(feature = "arango3_7"))]
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_post_create_document() {
     test_setup();
     let collection_name = "test_collection_create_document";
@@ -142,11 +138,7 @@ async fn test_post_create_document() {
 
 /// TODO need to use CI to validate this test
 #[cfg(any(feature = "arango3_7"))]
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_post_create_document_3_7() {
     test_setup();
     let collection_name = "test_collection_create_document_3_7";
@@ -324,11 +316,7 @@ async fn test_post_create_document_3_7() {
     coll.drop().await.expect("Should drop the collection");
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_get_read_document() {
     test_setup();
     let collection_name = "test_collection_read_document";
@@ -375,11 +363,7 @@ async fn test_get_read_document() {
     coll.drop().await.expect("Should drop the collection");
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_get_read_document_header() {
     test_setup();
     let collection_name = "test_collection_read_document_header";
@@ -452,11 +436,7 @@ async fn test_get_read_document_header() {
     coll.drop().await.expect("Should drop the collection");
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_patch_update_document() {
     test_setup();
     let collection_name = "test_collection_update_document";
@@ -504,7 +484,7 @@ async fn test_patch_update_document() {
 
     let result = update.unwrap();
     assert_eq!(
-        result.header().unwrap()._rev != _rev.to_string(),
+        result.header().unwrap()._rev != *_rev,
         true,
         "We should get a different revision after update"
     );
@@ -529,11 +509,7 @@ async fn test_patch_update_document() {
     // todo do more test for merge objects and stuff
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_post_replace_document() {
     test_setup();
     let collection_name = "test_collection_replace_document";
@@ -641,11 +617,7 @@ async fn test_post_replace_document() {
     // todo do more test
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_delete_remove_document() {
     test_setup();
     let collection_name = "test_collection_remove_document";
@@ -759,11 +731,7 @@ async fn test_delete_remove_document() {
     // todo do more test
 }
 
-#[maybe_async::test(
-    any(feature = "reqwest_blocking"),
-    async(any(feature = "reqwest_async"), tokio::test),
-    async(any(feature = "surf_async"), async_std::test)
-)]
+#[maybe_async::test(feature = "blocking", async(not(feature = "blocking"), tokio::test))]
 async fn test_document_deserialization() {
     use serde::{Deserialize, Serialize};
     #[derive(Debug, Default, Serialize, Deserialize)]